move system player updates out of media streamer, cleanup AVKit extensions and add deprecation notices
